refactor: extract shared SanitizedMarkdown component
- Create reusable SanitizedMarkdown component to eliminate code duplication
- Consolidate security-sensitive sanitization logic in one place
- Both IOModal and Playground now use the same implementation
- Reduces maintenance burden and ensures consistent security behavior
- Removes 105 net lines of duplicated code
- All 16 XSS security tests still passing
- Addresses code review feedback about duplicated security logic
